Develop

+++ src/main/java/com/wishlist/cst438project2/controller/AdminController.java
@@ -1,11 +1,10 @@
 package com.wishlist.cst438project2.controller;
 
 import com.wishlist.cst438project2.common.Constants;
-import com.wishlist.cst438project2.dto.ItemDTO;
-import com.wishlist.cst438project2.dto.SignInDTO;
-import com.wishlist.cst438project2.dto.SignUpDTO;
-import com.wishlist.cst438project2.dto.UserDTO;
+import com.wishlist.cst438project2.common.TokenManager;
+import com.wishlist.cst438project2.dto.*;
 import com.wishlist.cst438project2.exception.BadRequestException;
+import com.wishlist.cst438project2.exception.UnauthorizedException;
 import com.wishlist.cst438project2.service.AdminService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,6 +13,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 import java.util.Objects;
 
+@CrossOrigin
 @RestController
 @RequestMapping("/v1/admin")
 @Slf4j
@@ -21,6 +21,9 @@ public class AdminController {
 
     @Autowired
     private AdminService adminService;
+
+    @Autowired
+    private TokenManager tokenManager;
 
     @GetMapping("/users")
     public List<UserDTO> getAllUsers() {
@@ -46,11 +49,15 @@ public class AdminController {
      * @param username, of the user whose account is to be deleted.
      */
     @DeleteMapping("/deleteUser")
-    public String deleteUser(@RequestParam String username) {
+    public String deleteUser(@RequestHeader String accessToken, @RequestParam String username) {
 
         log.info("AdminController: Starting deleteUser");
 
         try{
+            UserTokenDTO userTokenDTO = tokenManager.getUser(accessToken);
+
+            if(Objects.isNull(userTokenDTO))
+                throw new UnauthorizedException(Constants.ERROR_INVALID_TOKEN);
 
             if(Objects.isNull(username) || username.isEmpty())
                 throw new BadRequestException();
@@ -73,11 +80,15 @@ public class AdminController {
      * @return user creation timestamp
      */
     @PostMapping("/createUser")
-    public String createUser(@RequestBody SignUpDTO signUpDTO) {
+    public String createUser(@RequestHeader String accessToken, @RequestBody SignUpDTO signUpDTO) {
 
         log.info("AdminController: Starting createUser");
 
         try {
+            UserTokenDTO userTokenDTO = tokenManager.getUser(accessToken);
+
+            if(Objects.isNull(userTokenDTO))
+                throw new UnauthorizedException(Constants.ERROR_INVALID_TOKEN);
 
             if (Objects.isNull(signUpDTO))
                 throw new BadRequestException();
@@ -104,11 +115,13 @@ public class AdminController {
             if (Objects.isNull(signInDTO))
                 throw new BadRequestException();
 
-            String responseTimestamp = adminService.login(signInDTO);
+            String accessToken = adminService.login(signInDTO);
+            if(Objects.isNull(accessToken))
+                throw new UnauthorizedException();
 
             log.info("AdminController: Exiting login");
 
-            return responseTimestamp;
+            return accessToken;
 
         } catch (Exception ex) {
             log.error(ex.getMessage(), ex);
@@ -121,9 +134,14 @@ public class AdminController {
      * returns item creation timestamp
      */
     @PostMapping("/createItem")
-    public String createItem(@RequestBody ItemDTO itemDTO) {
+    public String createItem(@RequestHeader String accessToken, @RequestBody ItemDTO itemDTO) {
         log.info("ItemController: Starting createItem");
         try {
+            UserTokenDTO userTokenDTO = tokenManager.getUser(accessToken);
+
+            if(Objects.isNull(userTokenDTO))
+                throw new UnauthorizedException(Constants.ERROR_INVALID_TOKEN);
+
             if (Objects.isNull(itemDTO) || (itemDTO.getName().isBlank() || Objects.isNull(itemDTO.getUserId()))) {
                 throw new BadRequestException();
             } else {
@@ -142,7 +160,13 @@ public class AdminController {
      * returns timestamp of successful deletion
      */
     @DeleteMapping("/removeItems")
-    public String removeItem(@RequestParam String itemName, @RequestParam int userId) {
+    public String removeItem(@RequestHeader String accessToken, @RequestParam String itemName, @RequestParam int userId) {
+
+        UserTokenDTO userTokenDTO = tokenManager.getUser(accessToken);
+
+        if(Objects.isNull(userTokenDTO))
+            throw new UnauthorizedException(Constants.ERROR_INVALID_TOKEN);
+
         log.info("ItemController: Starting removeItem");
         log.info("ItemController: removeItem:\n    name: {}\n    userId: {}", itemName, userId);
         return adminService.removeItem(itemName, userId);
